Rename src/Tlanislide.tsx -> src/Anipres.tsx
